Address most implicit conversion warnings.
- Harmonize variable, parameter, etc. data types more fully.
- Adjust system API usage accordingly:
-- For some string constants, use sizeof() - 1 in lieu of strlen,
   taking care to ensure they come from array variables.
-- Keep format specifiers for printf et al. in sync with argument types.
-- dbprrow: Account for the fact that printf always uses int for
   dynamic widths by substituting memchr + fwrite.
-- odbc/prepare_query.c (prepared_rpc): Substitute atoi for strtol
   (already used with no endptr) when populating TDSINT4.
- Cast away some remaining discrepancies where safe, in some cases by
  taking advantage of TDS_PUT_* macros.
- ctlib.h: Explicitly define _CS_CURS_TYPE_* in terms of
  TDS_CURSOR_STATE_* rather than duplicating their values.
- odbc_util.c: #include <stddef.h> for ptrdiff_t (now used in
  odbc_set_string_flag).
- tds_parse_conf_section: To legitimize a cast, tighten debug_flags'
  range check on systems where int is narrower than long.

+++ include/freetds/utils/des.h
@@ -25,7 +25,7 @@ void tds_des_encrypt(const DES_KEY * key, des_cblock block);
 #endif
 
 void tds_des_set_odd_parity(des_cblock key);
-int tds_des_ecb_encrypt(const void *plaintext, int len, DES_KEY * akey, uint8_t *output);
+int tds_des_ecb_encrypt(const void *plaintext, size_t len, DES_KEY * akey, uint8_t *output);
 
 #include <freetds/popvis.h>
